+++ src/features/auth/index.ts
@@ -1,0 +1,5 @@
+export { CheckEmail } from './ui/CheckEmail'
+export { CreateNewPasswordForm } from './ui/CreateNewPasswordForm'
+export { ForgotPasswordForm } from './ui/ForgotPasswordForm'
+export { SignInForm } from './ui/SignInForm'
+export { SignUpForm } from './ui/SignUpForm'
