+++ src/features/profile/index.ts
@@ -1,0 +1,1 @@
+export { PersonalInformation } from './ui/PersonalInformation'
